:sparkles: SpringBoot整合MyBatisPlus示例

+++ 004-springboot-mybatis-plus/springboot-mybatisplus-mysql/src/main/java/com/keqi/springbootmybatisplusmysql/common/AjaxEntity.java
@@ -57,7 +57,7 @@ public class AjaxEntity {
 	 *
 	 * @return ajaxEntity
 	 */
-	public static AjaxEntity list(int total, List<?> records) {
+	public static AjaxEntity list(long total, List<?> records) {
 		Map<String, Object> retMap = new HashMap<>();
 		retMap.put("total", total);
 		retMap.put("records", records);
